bug-15053: Replace size_t with uint64_t to allow support for 8gb files on 32-bit systems

+++ src/network_lib/include/jalop/jaln_network_types.h
@@ -263,7 +263,7 @@ struct jaln_payload_feeder {
 	 */
 	enum jal_status (*get_bytes)(const uint64_t offset,
 			   uint8_t * const buffer,
-			   uint32_t *size,
+			   uint64_t *size,
 			   void *feeder_data);
 };
